requirements.txt
# Netflix Mandate Wizard - Python Dependencies

# Core dependencies
pinecone-client==3.0.0
neo4j==5.14.0
sentence-transformers==2.2.2
openai==1.12.0
flask==3.0.0

# Data processing
numpy==1.24.3
pandas==2.0.3

# Optional (for development)
pytest==7.4.3

